fix(core): type the execution capability caps as integers
`maxIterations` and `maxExecutionTime` were `Option<f64>`, so a capabilities
document arriving with `"maxIterations": 10` was re-emitted as
`"maxIterations": 10.0`. A Rust service that reads an upstream agent's
capabilities and forwards them — a proxy, a registry, a router — silently
rewrote the payload it was only meant to relay, and a consumer parsing the
result into an integer field rejected a document its own SDK had produced.

The upstream Zod schema says `z.number()`, which is a JavaScript double and
therefore says nothing; the Python SDK says `Optional[int]`, and both an
iteration count and a millisecond budget are whole numbers wherever they are
produced. `timestamp` in this crate already resolves the same ambiguity the same
way. They are signed rather than unsigned because Python's `int` is: some
frameworks spell "no limit" as -1, and rejecting that would fail the whole
capabilities document over a value the reference implementations accept.

Adds tests/upstream_payloads.rs, which is what found this. The existing tests
build a Rust value, serialize it, and read it back — that proves the types are
self-consistent and nothing else, since a field wrong on the wire round-trips
perfectly through its own mistake. The new tests start from JSON transcribed
from the upstream schemas and assert byte-for-byte round-trips, plus a field
census pinning the exact key set each of the 33 events emits. The census found
no other discrepancy: names, optionality and camelCase are correct everywhere
else.

+++ crates/ag-ui-core/src/capabilities.rs
@@ -226,7 +226,17 @@ pub struct MultimodalCapabilities {
 }
 
 /// Execution controls and limits.
-#[derive(Clone, Copy, Debug, Default, PartialEq, Serialize, Deserialize)]
+///
+/// The two caps are integers, not floats. The upstream Zod schema types them as
+/// a bare `z.number()`, which is a JavaScript double — but the Python SDK
+/// declares them `Optional[int]`, and both an iteration count and a millisecond
+/// budget are whole numbers everywhere they are produced. Modelling them as
+/// `f64` would re-emit a received `"maxIterations": 10` as `10.0`, so a Rust
+/// proxy between two upstream implementations would silently rewrite the
+/// payload. They are signed because Python's `int` is: some frameworks spell
+/// "no limit" as `-1`, and rejecting that at parse time would fail the whole
+/// capabilities document over a value the reference implementations accept.
+#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 #[cfg_attr(feature = "schemars", derive(schemars::JsonSchema))]
 #[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
@@ -239,10 +249,10 @@ pub struct ExecutionCapabilities {
     pub sandboxed: Option<bool>,
     /// Cap on tool-call / reasoning iterations per run.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_iterations: Option<f64>,
+    pub max_iterations: Option<i64>,
     /// Wall-clock cap per run, in milliseconds.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub max_execution_time: Option<f64>,
+    pub max_execution_time: Option<i64>,
 }
 
 /// Human-in-the-loop support.
